+++ crates/tiangong-core/src/core/mod.rs
@@ -4,15 +4,13 @@
 //! CLI / GUI / Server / Connector 统一通过 TiangongCore 运行。
 
 use std::sync::Arc;
-use std::sync::atomic::Ordering;
 use std::sync::mpsc::Sender;
-use tokio::sync::mpsc as tokio_mpsc;
 use typed_builder::TypedBuilder;
 
 use crate::core_config::{CoreConfig, CoreConfigProvider};
 use crate::model::SingleProviderClient;
-use crate::react::message::{AcceptedUserMessage, accept_user_message};
-use crate::session::{MessageRole, Session};
+use crate::react::turn::run_turn;
+use crate::session::Session;
 use tiangong_types::StreamEvent;
 
 pub mod command;
@@ -48,9 +46,6 @@ pub struct TiangongCore {
     stream_tx: Sender<StreamEvent>,
     /// 进程内插件(会话级,跨 turn)。
     plugins: Vec<Arc<dyn Plugin>>,
-    /// on_session_ready 是否已触发(跨 turn)。
-    #[builder(default = std::sync::Arc::new(std::sync::atomic::AtomicBool::new(false)))]
-    session_ready_fired: Arc<std::sync::atomic::AtomicBool>,
 }
 
 impl TiangongCore {
@@ -132,24 +127,26 @@ impl crate::agent_input::AgentInput for TiangongCore {
                     while crate::shared_runtime::is_running(&self.session_id) {
                         if std::time::Instant::now() > deadline {
                             tracing::warn!(session_id = %self.session_id, "等待上一轮 turn 结束超时");
-                            break;
+                            return Err(CoreError::WorkerPanicked);
                         }
                         std::thread::sleep(std::time::Duration::from_millis(50));
                     }
                 }
 
                 let session_id = self.session_id.clone();
-                let session = Session::load_from_storage(&self.storage_root, &session_id).map_err(
-                    |error| {
+                let mut session = Session::load_from_storage(&self.storage_root, &session_id)
+                    .map_err(|error| {
                         tracing::warn!(%error, %session_id, "加载本轮 Session 失败");
                         CoreError::WorkerStopped
-                    },
-                )?;
+                    })?;
+                session.append_prepared_user_message_with_id(
+                    message_id.unwrap_or(scru128::new_string()),
+                    prepared,
+                );
                 let trust_mode = *self.trust_mode.lock().unwrap_or_else(|p| p.into_inner());
                 let config = (*self.config.snapshot()).clone();
                 let storage_root = self.storage_root.clone();
                 let plugins = self.plugins.clone();
-                let session_ready_fired = self.session_ready_fired.clone();
                 // stream_tx 直接用 core 的通道(发 StreamEvent 到 app 层)
                 let stream_tx = self.stream_tx.clone();
 
@@ -173,7 +170,7 @@ impl crate::agent_input::AgentInput for TiangongCore {
                 let usage_sink = Arc::new(crate::core::plugin::TurnUsageSink::new());
                 let prepared_plugins =
                     crate::core::plugin::prepare_plugins(&plugins, &config, trust_mode, &session);
-                let mut ctx = crate::turn_context::TurnContext::builder()
+                let ctx = crate::turn_context::TurnContext::builder()
                     .client(client)
                     .session(session)
                     .stream_tx(stream_tx.clone())
@@ -187,21 +184,11 @@ impl crate::agent_input::AgentInput for TiangongCore {
                     .tools(prepared_plugins.tools)
                     .build();
 
-                let message_id = message_id.unwrap_or_else(|| scru128::new().to_string());
-                let mut session = std::mem::replace(&mut ctx.session, Session::new("placeholder"));
-                let accepted =
-                    accept_user_message(&mut session, &ctx, Some(message_id), prepared, true)
-                        .map_err(|error| {
-                            tracing::warn!(%error, "持久化本轮用户消息失败");
-                            CoreError::WorkerStopped
-                        })?;
-                ctx.session = session;
-
                 crate::shared_runtime::spawn_turn(ctx, move |mut ctx, cmd_rx| {
-                    if !session_ready_fired.swap(true, Ordering::AcqRel) {
-                        for plugin in &ctx.plugins {
-                            plugin.on_session_ready(&mut ctx.session);
-                        }
+                    // 每轮都触发 on_session_ready:插件自行保证幂等性(如 index 扫描的 last_scanned
+                    // 标志、memory 的 session_id 已设置检查、agent-team 的 manifest 已存在检查)。
+                    for plugin in &ctx.plugins {
+                        plugin.on_session_ready(&mut ctx.session);
                     }
                     let prompt_sections = ctx
                         .plugins
@@ -215,7 +202,7 @@ impl crate::agent_input::AgentInput for TiangongCore {
                         tracing::warn!(%error, "持久化本轮系统提示失败");
                         CoreError::WorkerStopped
                     })?;
-                    Ok(run_turn(ctx, cmd_rx, accepted))
+                    Ok(run_turn(ctx, cmd_rx))
                 })
             }
             AgentInputKind::Tool(tool) => self.send_cmd(Command::InjectTool {
@@ -252,123 +239,6 @@ impl Drop for TiangongCore {
     }
 }
 
-/// turn task:执行 TurnContext 中的 turn。
-///
-/// TurnContext(session + 用户消息已注入)在 deliver 中构建,传入此处执行。
-/// turn 结束后 session 落盘,task 退出。
-async fn run_turn(
-    mut ctx: crate::turn_context::TurnContext,
-    mut cmd_rx: tokio_mpsc::UnboundedReceiver<Command>,
-    accepted: AcceptedUserMessage,
-) {
-    let stream_tx = ctx.stream_tx.clone();
-    let turn_start_idx = accepted.turn_start_idx;
-    let user_msg_id = accepted.message_id;
-    let prepared = accepted.prepared;
-    let turn_started = std::time::Instant::now();
-    let turn_start_cwd = ctx.session.cwd.clone();
-
-    for plugin in &ctx.plugins {
-        plugin.on_turn_started(&mut ctx.session, turn_start_idx);
-    }
-
-    let mut session = std::mem::replace(&mut ctx.session, Session::new("placeholder"));
-    let usage = ctx
-        .execute_turn(&mut session, Some((&user_msg_id, &prepared)), &mut cmd_rx)
-        .await;
-    ctx.session = session;
-
-    // execute_turn 返回的 usage 直接应用到 session
-    ctx.session.token_usage.accumulate(&usage);
-
-    let mut terminal = StreamEvent::Done { usage: None };
-
-    if ctx.session.cwd != turn_start_cwd {
-        let workspace_path = std::path::PathBuf::from(&ctx.session.cwd);
-        let workspace = workspace_path.is_dir().then_some(workspace_path.as_path());
-        for plugin in &ctx.plugins {
-            plugin.set_workspace(workspace);
-        }
-    }
-
-    let interrupted_tools = {
-        let mut session = std::mem::replace(&mut ctx.session, Session::new("placeholder"));
-        let result = crate::react::message::close_unfinished_tool_calls_for_turn(&mut session);
-        ctx.session = session;
-        result
-    };
-    if !interrupted_tools.is_empty() {
-        for (tool_call_id, tool_name, output) in interrupted_tools {
-            let _ = stream_tx.send(StreamEvent::ToolResult {
-                name: tool_name,
-                tool_call_id: Some(tool_call_id),
-                ok: false,
-                output,
-                full_output: None,
-                duration_ms: None,
-            });
-        }
-        terminal = StreamEvent::Error {
-            message: "本轮仍有未完成的工具调用，已安全中断".to_string(),
-        };
-    }
-
-    {
-        let mut session = std::mem::replace(&mut ctx.session, Session::new("placeholder"));
-        crate::react::message::flush_deferred_tool_injections(&mut session, &ctx);
-        ctx.session = session;
-    }
-
-    let elapsed_ms = turn_started.elapsed().as_millis() as u64;
-    let mut status = TurnOutcome::from_terminal(&terminal).status;
-
-    if status == tiangong_types::TurnStatus::Cancelled {
-        for plugin in &ctx.plugins {
-            plugin.on_cancel(&mut ctx.session).await;
-        }
-    }
-
-    let mut user_msg_updated = false;
-    if let Some(msg) = ctx
-        .session
-        .messages
-        .iter_mut()
-        .find(|m| m.id == user_msg_id && m.role == MessageRole::User)
-    {
-        msg.set_turn_result(elapsed_ms, status);
-        user_msg_updated = true;
-    }
-    for plugin in &ctx.plugins {
-        plugin.on_turn_finished(&mut ctx.session, turn_start_idx);
-    }
-
-    ctx.session.clear_transient_content();
-
-    if let Err(error) = ctx.session.try_persist_to_disk() {
-        terminal = StreamEvent::Error {
-            message: format!("最终会话持久化失败：{error}"),
-        };
-        status = tiangong_types::TurnStatus::Failed;
-        if let Some(msg) = ctx
-            .session
-            .messages
-            .iter_mut()
-            .find(|m| m.id == user_msg_id && m.role == MessageRole::User)
-        {
-            msg.set_turn_result(elapsed_ms, status);
-        }
-        let _ = ctx.session.try_persist_to_disk();
-    }
-    if user_msg_updated {
-        crate::react::message::emit_session_message_upsert(&ctx.session, &ctx, &user_msg_id);
-    }
-
-    // 直接发送终态事件(无 forwarder 屏障)
-    let _ = stream_tx.send(terminal);
-
-    drop(ctx);
-}
-
 pub(crate) fn reset_context_for_session(
     session: &mut Session,
     ctx: &crate::turn_context::TurnContext,
@@ -388,37 +258,6 @@ pub(crate) fn reset_context_for_session(
         remaining_messages: 0,
     });
     session.persist_to_disk();
-}
-
-/// 转发线程捕获的单个对话轮次终态。
-///
-/// `status` 由终态事件推导：`Done` → Success；`Error` 文案含「取消/cancel/abort」
-/// 时为 Cancelled，否则为 Failed。run_turn 据此把 `status` 与执行时长
-/// 写入用户消息，供前端（含历史会话）展示。
-struct TurnOutcome {
-    status: tiangong_types::TurnStatus,
-}
-
-impl TurnOutcome {
-    fn from_terminal(event: &StreamEvent) -> Self {
-        let status = match event {
-            StreamEvent::Done { .. } => tiangong_types::TurnStatus::Success,
-            StreamEvent::Error { message } => {
-                let lower = message.to_lowercase();
-                if lower.contains("取消")
-                    || lower.contains("cancel")
-                    || lower.contains("abort")
-                    || lower.contains("中断")
-                {
-                    tiangong_types::TurnStatus::Cancelled
-                } else {
-                    tiangong_types::TurnStatus::Failed
-                }
-            }
-            _ => tiangong_types::TurnStatus::Failed,
-        };
-        Self { status }
-    }
 }
 
 #[cfg(test)]
